fix basic layout
