+++ frontend/src/components/layout/BasicLayout/BasicLayout.tsx
@@ -1,0 +1,14 @@
+import React, { FC, ReactNode } from "react";
+import { Breadcrumbs } from "../../breadcrumbs/Breadcrumbs";
+import classes from "./BasicLayout.module.scss";
+
+export const BasicLayout: FC<{ children: ReactNode }> = (props) => {
+  return (
+    <div className={classes.root}>
+      <div className={classes.header}>
+        <Breadcrumbs className={classes.breadcrumbs} />
+      </div>
+      {props.children}
+    </div>
+  );
+};
